feat(tests): move products controller tests to user products controller tests
